fix: moves tests for `test` module to a separate file
